+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix a few bugs
 - Update to 1.21.5, drop 1.21.4
+- Add Tablist background customization (closes #120)
 
 ### 3.1.2